Refactor: cleanup add vault static data in dedicated component
