+++ front/components/vaults/VaultResourcesList.tsx
@@ -1,12 +1,4 @@
-import {
-  Button,
-  Chip,
-  DataTable,
-  PlusIcon,
-  Popup,
-  Searchbar,
-  Spinner,
-} from "@dust-tt/sparkle";
+import { Chip, DataTable, Searchbar, Spinner } from "@dust-tt/sparkle";
 import type {
   ConnectorType,
   DataSourceViewCategory,
@@ -17,7 +9,6 @@ import type {
 } from "@dust-tt/types";
 import type { CellContext } from "@tanstack/react-table";
 import { FolderIcon } from "lucide-react";
-import { useRouter } from "next/router";
 import type { ComponentType } from "react";
 import { useRef } from "react";
 import { useState } from "react";
@@ -25,9 +16,7 @@ import * as React from "react";
 
 import ConnectorSyncingChip from "@app/components/data_source/DataSourceSyncChip";
 import { EditVaultManagedDataSourcesViews } from "@app/components/vaults/EditVaultManagedDatasourcesViews";
-import VaultCreateFolderModal from "@app/components/vaults/VaultCreateFolderModal";
-import VaultCreateWebsiteModal from "@app/components/vaults/VaultCreateWebsiteModal";
-import { useSubmitFunction } from "@app/lib/client/utils";
+import { EditVaultStaticDataSourcesViews } from "@app/components/vaults/EditVaultStaticDatasourcesViews";
 import {
   CONNECTOR_CONFIGURATIONS,
   getDataSourceNameFromView,
@@ -125,13 +114,6 @@ export const VaultResourcesList = ({
   category,
   onSelect,
 }: VaultResourcesListProps) => {
-  const router = useRouter();
-
-  const [showDatasourceLimitPopup, setShowDatasourceLimitPopup] =
-    useState(false);
-  const [showAddFolderModal, setShowAddFolderModal] = useState(false);
-  const [showAddWebsiteModal, setShowAddWebsiteModal] = useState(false);
-
   const [dataSourceSearch, setDataSourceSearch] = useState<string>("");
 
   const { dataSources, isDataSourcesLoading } = useDataSources(owner);
@@ -145,21 +127,6 @@ export const VaultResourcesList = ({
       vaultId: vault.sId,
       category: category,
     });
-
-  const { submit: handleCreateStaticDataSource } = useSubmitFunction(
-    async (type: "folder" | "website") => {
-      if (
-        plan.limits.dataSources.count != -1 &&
-        dataSources.length >= plan.limits.dataSources.count
-      ) {
-        setShowDatasourceLimitPopup(true);
-      } else if (type === "folder") {
-        setShowAddFolderModal(true);
-      } else if (type === "website") {
-        setShowAddWebsiteModal(true);
-      }
-    }
-  );
 
   const rows: RowData[] =
     vaultDataSourceViews?.map((r) => ({
@@ -187,35 +154,6 @@ export const VaultResourcesList = ({
 
   return (
     <>
-      <Popup
-        show={showDatasourceLimitPopup}
-        chipLabel={`${plan.name} plan`}
-        description={`You have reached the limit of data sources (${plan.limits.dataSources.count} data sources). Upgrade your plan for unlimited datasources.`}
-        buttonLabel="Check Dust plans"
-        buttonClick={() => {
-          void router.push(`/w/${owner.sId}/subscription`);
-        }}
-        onClose={() => {
-          setShowDatasourceLimitPopup(false);
-        }}
-        className="absolute bottom-8 right-0"
-      />
-      <VaultCreateFolderModal
-        isOpen={showAddFolderModal}
-        setOpen={(isOpen) => {
-          setShowAddFolderModal(isOpen);
-        }}
-        owner={owner}
-        vault={vault}
-        dataSources={dataSources}
-      />
-      <VaultCreateWebsiteModal
-        isOpen={showAddWebsiteModal}
-        setOpen={(isOpen) => {
-          setShowAddWebsiteModal(isOpen);
-        }}
-        owner={owner}
-      />
       <div
         className={classNames(
           "flex gap-2",
@@ -242,22 +180,13 @@ export const VaultResourcesList = ({
             systemVault={systemVault}
           />
         )}
-        {category === "folder" && (
-          <Button
-            label="Add folder"
-            onClick={async () => {
-              await handleCreateStaticDataSource("folder");
-            }}
-            icon={PlusIcon}
-          />
-        )}
-        {category === "website" && (
-          <Button
-            label="Add site"
-            onClick={async () => {
-              await handleCreateStaticDataSource("website");
-            }}
-            icon={PlusIcon}
+        {(category === "folder" || category === "website") && (
+          <EditVaultStaticDataSourcesViews
+            owner={owner}
+            plan={plan}
+            vault={vault}
+            category={category}
+            dataSources={dataSources}
           />
         )}
       </div>
